app/requirements.txt
wheel
fastapi
motor
uvicorn
python-dotenv
pydantic
bcrypt
python-jose